update-entrega

+++ scripts/tasks.js
@@ -1,39 +1,21 @@
-const urlUsuario = "https://todo-api.ctd.academy/v1/users/getMe";
-const token = localStorage.getItem("token");
-const novaTarefa = document.querySelector("#novaTarefa");
-
-const btnSubmitRef = document.querySelector("#btnSubmit");
-
-// ----------------------------------------------------------------------------------------------
-// CHECAR TOKEN
-// ----------------------------------------------------------------------------------------------
-
-const tokenCheck = localStorage.getItem("token");
-function checkToken() {
-  if (tokenCheck === null) {
-    location.replace("./index.html");
-  }
-}
-
-// ------------------- FUNÇÃO VALIDAR INPUT--------------------------------
-const novaTarefaRef = document.querySelector("#novaTarefa")
-validateInputTask()
-function validateInputTask(){
-  const novaTarefa = novaTarefaRef.value.trim();
-
-  if (novaTarefa.length <4){
-    novaTarefaRef.classList.add("error");
-    return false;
-  }else {
-      novaTarefaRef.classList.remove("error");
-    return true;
-  }
-}
-checkToken();
-// ----------------------------------------------------------------------------------------------
-//         1-BOTAO SAIR
-// ----------------------------------------------------------------------------------------------
 document.addEventListener("DOMContentLoaded", () => {
+
+  const urlUsuario = "https://todo-api.ctd.academy/v1/users/getMe";
+  const token = localStorage.getItem("token");
+  const novaTarefa = document.querySelector("#novaTarefa");
+
+  const btnSubmitRef = document.querySelector("#btnSubmit");
+
+
+  const tokenCheck = localStorage.getItem("token");
+  function checkToken() {
+    if (tokenCheck === null) {
+      location.replace("./index.html");
+    }
+  }
+
+  checkToken();
+ 
   const closeButtonRef = document.querySelector("#closeApp");
   function logout() {
     Swal.fire({
@@ -52,9 +34,7 @@ document.addEventListener("DOMContentLoaded", () => {
     });
   }
 
-  // ----------------------------------------------------------------------------------------------
-  //              2-Nome de usuario get
-  // ----------------------------------------------------------------------------------------------
+ 
   obterNomeUsuario();
   function obterNomeUsuario() {
     const settings = {
@@ -74,15 +54,10 @@ document.addEventListener("DOMContentLoaded", () => {
         location.replace("./index.html");
       });
   }
-  /* -------------------------------------------------------------------------- */
-  /*                 3-Buscar  tarefas  get          */
-  /* -------------------------------------------------------------------------- */
-  getTasks();
-
+  
   const buscarTarefas = "https://todo-api.ctd.academy/v1/tasks";
-
   function getTasks() {
-    var requestConfig = {
+    const requestConfig = {
       method: "GET",
       headers: {
         authorization: token,
@@ -99,56 +74,93 @@ document.addEventListener("DOMContentLoaded", () => {
       }
     );
   }
-  // ----------------------------------------------------------------------------------------------
-  //                    4 Criar Novas tarefas POST
-  // ----------------------------------------------------------------------------------------------
+
+  getTasks();
+
+ 
+  const novaTarefaRef = document.querySelector("#novaTarefa");
+  function validateInputTask() {
+    const novaTarefa = novaTarefaRef.value.trim();
+
+    if (novaTarefa.length < 4) {
+      novaTarefaRef.classList.add("error");
+      return false;
+    } else {
+      novaTarefaRef.classList.remove("error");
+
+      return true;
+    }
+  }
+  novaTarefaRef.addEventListener("keyup", validateInputTask);
+
+
+  function removeErrorClass() {
+    novaTarefaRef.classList.remove("error");
+  }
+  novaTarefaRef.addEventListener("blur", removeErrorClass);
+
+ 
   function criarTarefa(event) {
     event.preventDefault();
 
-    const tarefas = {
-      description: novaTarefa.value,
-      completed: false,
-    };
+    if (validateInputTask()) {
+      const tarefas = {
+        description: novaTarefa.value,
+        completed: false,
+      };
 
-    const requestHeaders = {
-      Accept: "application/json",
-      "Content-Type": "application/json",
-      Authorization: token,
-    };
+      const requestHeaders = {
+        Accept: "application/json",
+        "Content-Type": "application/json",
+        Authorization: token,
+      };
 
-    const requestConfig = {
-      method: "POST",
-      body: JSON.stringify(tarefas),
-      headers: requestHeaders,
-    };
-
-    fetch(buscarTarefas, requestConfig)
-      .then((response) => response.json())
-      .then((tarefa) => {
-        getTasks();
-    
-
-        novaTarefa.value = "";
-      });
+      const requestConfig = {
+        method: "POST",
+        body: JSON.stringify(tarefas),
+        headers: requestHeaders,
+      };
+      fetch(buscarTarefas, requestConfig)
+        .then((response) => response.json())
+        .then(() => {
+          getTasks();
+          novaTarefa.value = "";
+        });
+    }
   }
-  // ----------------------------------------------------------------------------------------------
-  // --------------- OBTEN AS TAREFAS DA API  [GET]
-  // ----------------------------------------------------------------------------------------------
+  
+  function orderByDate(a, b) {
+    const dateRecent = new Date(a.createdAt);
+    const dateOld = new Date(b.createdAt);
+
+    if (dateRecent > dateOld) {
+      return -1;
+    } else if (dateRecent < dateOld) {
+      return 1;
+    } else {
+      return 0;
+    }
+  }
+
+ 
   function getTasksOnApi(tarefas) {
     const tarefasPendentes = document.querySelector(".tarefas-pendentes");
     const tarefasTerminadas = document.querySelector(".tarefas-terminadas");
     tarefasPendentes.innerHTML = "";
     tarefasTerminadas.innerHTML = "";
 
+    tarefas.sort(orderByDate);
+
     if (tarefas && Array.isArray(tarefas)) {
       tarefas.forEach((tarefa) => {
-        
         const createdAtDate = new Date(tarefa.createdAt);
-        const createAtFormted = new Intl.DateTimeFormat("pt-BR").format(createdAtDate);
+        const createAtFormted = new Intl.DateTimeFormat("pt-BR").format(
+          createdAtDate
+        );
 
         if (tarefa.completed) {
           tarefasTerminadas.innerHTML += `
-         <li class="tarefa">
+         <li class="tarefa"  >
          <div class="not-done finish" id="${tarefa.id}"></div>
          <div class="descricao">
            <p class="nome">${tarefa.description}</p>
@@ -173,18 +185,15 @@ document.addEventListener("DOMContentLoaded", () => {
     }
   }
 
-  // ----------------------------------------------------------------------------------------------
-  // --------------- EVENT LISTENER PPARA TROCAR ESTADO DA TAREFA
-  // --------------------------------------------------------------------------------------------
+
   document.addEventListener("click", (event) => {
     if (event.target.classList.contains("unfinished")) {
       const taskId = event.target.attributes.id.value;
       atualizarTarefa(taskId);
     }
   });
-  // ----------------------------------------------------------------------------------------------
-  // --------------- FUNÇÃO PARA TROCAR ESTADO DA TAREFA
-  // --------------------------------------------------------------------------------------------
+
+
   function atualizarTarefa(id) {
     let tarefaObj = {
       completed: true,
@@ -206,51 +215,57 @@ document.addEventListener("DOMContentLoaded", () => {
     fetch(`https://todo-api.ctd.academy/v1/tasks/${id}`, requestConfig).then(
       (response) => {
         if (response.ok) {
-          // console.log(`Tarefa ${id} atualizada com sucesso!`);
-          console.log(response)
+        
+
           getTasks();
         } else {
-          // console.error("Houve um erro ao atualizar a tarefa (requisição PUT)");
+          console.error("Houve um erro ao atualizar a tarefa (requisição PUT)");
         }
       }
     );
   }
 
-  // ===========================================================
-  // 6 - EVENT LISTENER PARA DELETAR TASKS POR ID
-  // ===========================================================
+
   document.addEventListener("click", (event) => {
     if (event.target.classList.contains("finish")) {
       const taskId = event.target.attributes.id.value;
       deletarTarefa(taskId);
     }
   });
-  // ===========================================================
-  // Função que deleta uma Task pelo ID
-  // ===========================================================
+
   function deletarTarefa(taskId) {
-    const requestHeaders = {
-      Accept: "application/json",
-      "Content-Type": "application/json",
-      Authorization: token,
-    };
+    Swal.fire({
+      title: "Deletar a Tarefa?",
+      icon: "question",
+      showCancelButton: true,
+      confirmButtonColor: "#7898FF;",
+      cancelButtonColor: "#8E64C5",
+      confirmButtonText: "Sim delete  ",
+      cancelButtonText: "Mudei de ideia",
+    }).then((result) => {
+      if (result.isConfirmed) {
+        const requestHeaders = {
+          Accept: "application/json",
+          "Content-Type": "application/json",
+          Authorization: token,
+        };
 
-    const requestConfig = {
-      method: "DELETE",
-      headers: requestHeaders,
-    };
+        const requestConfig = {
+          method: "DELETE",
+          headers: requestHeaders,
+        };
 
-    fetch(`${buscarTarefas}/${taskId}`, requestConfig)
-      .then((response) => {
-        if (response.ok) {
-          // console.log(`Tarefa ${taskId} deletada com sucesso!`);
-
-          getTasks();
-        }
-      })
-      .catch((error) => {
-        // console.error("Houve um erro ao deletar a tarefa: ", error);
-      });
+        fetch(`${buscarTarefas}/${taskId}`, requestConfig).then((response) => {
+          if (response.ok) {
+            getTasks();
+            Swal.fire({
+              title: "Tarefa deletada com sucesso!",
+              icon: "success",
+            });
+          } 
+        });
+      }
+    });
   }
 
   btnSubmitRef.addEventListener("click", criarTarefa);
